+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - Explicitly rescue NameError in Cache#presenter_classes_for
 - Add Oprah::TestHelpers
+- Delegate #to_s and #inspect to the presented object
 
 0.1.1
 -----